requirements.txt
# ✅ Compatible with Python 3.12 and LangChain 0.3.26

# Core AI / LLM
langchain==0.3.26
langchain-community==0.3.26
langchain-huggingface==0.3.0
huggingface_hub==0.30.2      # ✅ required minimum version for langchain-huggingface
openai==1.30.1

# Routing / planning
langgraph==0.5.0

# Vector store / embeddings
faiss-cpu==1.8.0
sentence-transformers==2.7.0

# Frontend and utilities
streamlit==1.35.0
streamlit-folium==0.25.0
folium==0.16.0
pandas==2.2.2
python-dotenv==1.0.1
requests==2.31.0
